refactor: change helper module name
Capitalise the H so that it is in keeping with other modules within the dir.

Refs: N/A
